move move function to player repo instead of map repo
