+++ src/main/java/com/nscharrenberg/um/multiagentsurveillance/headless/contracts/repositories/IPlayerRepository.java
@@ -1,11 +1,38 @@
 package com.nscharrenberg.um.multiagentsurveillance.headless.contracts.repositories;
 
+import com.nscharrenberg.um.multiagentsurveillance.headless.exceptions.CollisionException;
+import com.nscharrenberg.um.multiagentsurveillance.headless.exceptions.InvalidTileException;
+import com.nscharrenberg.um.multiagentsurveillance.headless.exceptions.ItemAlreadyOnTileException;
+import com.nscharrenberg.um.multiagentsurveillance.headless.exceptions.ItemNotOnTileException;
+import com.nscharrenberg.um.multiagentsurveillance.headless.models.Angle;
 import com.nscharrenberg.um.multiagentsurveillance.headless.models.Guard;
 import com.nscharrenberg.um.multiagentsurveillance.headless.models.Intruder;
+import com.nscharrenberg.um.multiagentsurveillance.headless.models.Player;
 
 import java.util.List;
 
 public interface IPlayerRepository {
+
+    /**
+     * Rotate or Move a player to a new tile
+     * @param player - the player to move
+     * @param direction - the direction to move to
+     * @throws CollisionException - Thrown when the player is about to move into an inactive Tile
+     * @throws InvalidTileException - Thrown when the player is trying to move to an invalid tile
+     * @throws ItemNotOnTileException - Thrown when the player is not on the tile (Should not happen)
+     * @throws ItemAlreadyOnTileException - Thrown when the player is already on the tile its trying to move to (should not happen)
+     */
+    void move(Player player, Angle direction) throws CollisionException, InvalidTileException, ItemNotOnTileException, ItemAlreadyOnTileException;
+
+    /**
+     * Validates whether the move the player wants to make is a valid move
+     * Same logic as the `move` function but not actually making the move and instead giving a valid/invalid indication
+     * @param player - the player to move
+     * @param direction - the direction to move to
+     * @return whether it is a valid move or not
+     */
+    boolean isLegalMove(Player player, Angle direction);
+
     List<Intruder> getIntruders();
 
     void setIntruders(List<Intruder> intruders);
